+++ examples/key_values_sorted.rs
@@ -10,7 +10,7 @@ use std::{
 
 use bevy::prelude::*;
 use bevy_cosmic_edit::{CosmicBackgroundColor, CosmicWrap, CursorColor, MaxLines};
-use haalka::{prelude::*, text_input::FocusedTextInput};
+use haalka::{prelude::*, text_input::FocusedTextInput, viewport_mutable::MutableViewport};
 
 fn main() {
     App::new()
@@ -504,42 +504,48 @@ fn escaper(keys: Res<ButtonInput<KeyCode>>, mut commands: Commands) {
 // on focus change, check if the focused element is in view, if not, scroll to it
 fn focus_scroller(
     focused_text_input_option: Option<Res<FocusedTextInput>>,
-    data_query: Query<(&Node, &GlobalTransform, &Parent, &mut Style)>,
+    data_query: Query<(&Node, &GlobalTransform, &mut Style)>,
+    parents: Query<&Parent>,
+    mutable_viewports: Query<&MutableViewport>,
 ) {
     if let Some(focused_text_input) = focused_text_input_option.as_deref().map(Deref::deref).copied() {
-        if let Ok((child_node, child_transform, child_parent, _)) = data_query.get(focused_text_input) {
-            // TODO: what is this node ?
-            if let Ok((_, _, child_parent, _)) = data_query.get(child_parent.get()) {
-                if let Ok((scrollable_node, scrollable_transform, scrollable_container, _)) =
-                    data_query.get(child_parent.get())
-                {
-                    if let Ok((
-                        scrollable_container_node,
-                        scrollable_container_transform,
-                        _,
-                        scrollable_container_style,
-                    )) = data_query.get(scrollable_container.get())
+        if let Ok((text_input_node, text_input_transform, _)) = data_query.get(focused_text_input) {
+            for parent in parents.iter_ancestors(focused_text_input) {
+                if mutable_viewports.contains(parent) {
+                    if let Ok((scene_node, scene_transform, scene_style)) =
+                        data_query.get(parent)
                     {
-                        let child_rect = child_node.logical_rect(child_transform);
-                        let scrollable_rect = scrollable_node.logical_rect(scrollable_transform);
-                        let scrollable_container_rect =
-                            scrollable_container_node.logical_rect(scrollable_container_transform);
-                        let scrolled_option = match scrollable_container_style.top {
-                            Val::Px(top) => Some(top),
-                            Val::Auto => Some(0.0),
-                            _ => None,
-                        };
-                        if let Some(scrolled) = scrolled_option {
-                            let container_base = scrollable_container_rect.min.y - scrolled;
-                            let child_offset = child_rect.min.y - scrolled - container_base;
-                            // TODO: is there a simpler/ more general way to check for node visibility ?
-                            if child_offset + INPUT_HEIGHT - scrolled > scrollable_container_rect.height() {
-                                SCROLL_POSITION.set(
-                                    scrollable_rect.min.y - child_rect.min.y + scrollable_container_rect.height()
-                                        - INPUT_HEIGHT,
-                                );
-                            } else if child_offset < scrolled {
-                                SCROLL_POSITION.set(scrollable_rect.min.y - child_rect.min.y);
+                        if let Some((
+                            viewport_node,
+                            viewport_transform,
+                            viewport_style,
+                        )) = parents.get(parent).ok().and_then(|parent| data_query.get(parent.get()).ok())
+                        {
+                            let text_input_rect = text_input_node.logical_rect(text_input_transform);
+                            let scene_rect = scene_node.logical_rect(scene_transform);
+                            let viewport_rect =
+                                viewport_node.logical_rect(viewport_transform);
+                            let scrolled_option = match scene_style.top {
+                                Val::Px(top) => Some(top),
+                                Val::Auto => Some(0.0),
+                                _ => None,
+                            };
+                            println!(
+                                "text_input_rect: {:?}\nscene_rect: {:?}\nviewport_rect: {:?}\nscrolled: {:?}",
+                                text_input_rect, scene_rect, viewport_rect, scrolled_option
+                            );
+                            if let Some(scrolled) = scrolled_option {
+                                let container_base = viewport_rect.min.y - scrolled;
+                                let child_offset = text_input_rect.min.y - scrolled - container_base;
+                                // TODO: is there a simpler/ more general way to check for node visibility ?
+                                if child_offset + INPUT_HEIGHT - scrolled > viewport_rect.height() {
+                                    SCROLL_POSITION.set(
+                                        scene_rect.min.y - text_input_rect.min.y + viewport_rect.height()
+                                            - INPUT_HEIGHT,
+                                    );
+                                } else if child_offset < scrolled {
+                                    SCROLL_POSITION.set(scene_rect.min.y - text_input_rect.min.y);
+                                }
                             }
                         }
                     }