+++ src/viewport_mutable.rs
@@ -1,7 +1,14 @@
-use super::raw::{DeferredUpdaterAppendDirection, RawElWrapper, RawHaalkaEl};
-use bevy::prelude::*;
-use futures_signals::signal::Signal;
+use crate::raw::{observe, register_system};
 
+use super::{
+    raw::{utils::remove_system_holder_on_remove, DeferredUpdaterAppendDirection, RawElWrapper, RawHaalkaEl},
+    utils::clone,
+};
+use apply::Apply;
+use bevy::{prelude::*, ui::ui_layout_system};
+use futures_signals::signal::{Mutable, Signal};
+
+#[derive(Clone, Copy)]
 pub enum LimitToBody {
     Horizontal,
     Vertical,
@@ -10,21 +17,50 @@ pub enum LimitToBody {
 
 // can also be used to query for mutable viewports
 #[derive(Component)]
-pub struct MutableViewportSettings {
+pub struct MutableViewport {
     limit_to_body: Option<LimitToBody>,
+    scene: Scene,
+    viewport: Viewport,
 }
 
-// pub struct Viewport {
-//     pub x: f32,
-//     pub y: f32,
-//     width: f32,
-//     height: f32,
-// }
+impl MutableViewport {
+    pub fn new(limit_to_body: Option<LimitToBody>) -> Self {
+        Self {
+            limit_to_body,
+            scene: default(),
+            viewport: default(),
+        }
+    }
 
-// pub struct Scene {
-//     pub width: f32,
-//     pub height: f32,
-// }
+    pub fn scene(&self) -> Scene {
+        self.scene
+    }
+
+    pub fn viewport(&self) -> Viewport {
+        self.viewport
+    }
+}
+
+#[derive(Component)]
+pub struct ViewportMarker;
+
+#[derive(Clone, Copy, Default, Debug)]
+pub struct Viewport {
+    /// Horizontal offset.
+    pub x: f32,
+    /// Vertical offset.
+    pub y: f32,
+    #[allow(missing_docs)]
+    pub width: f32,
+    #[allow(missing_docs)]
+    pub height: f32,
+}
+
+#[derive(Clone, Copy, Default, Debug)]
+pub struct Scene {
+    pub width: f32,
+    pub height: f32,
+}
 
 #[derive(Default, Event)]
 pub struct ViewportMutation {
@@ -52,6 +88,9 @@ impl ViewportMutation {
     }
 }
 
+#[derive(Component)]
+pub struct OnViewportLocationChange;
+
 /// Enables the management of a limited visible window (viewport) onto the body of an element.
 /// CRITICALLY NOTE that methods expecting viewport mutability will not function without calling
 /// [`.mutable_viewport(...)`](ViewportMutable::mutable_viewport).
@@ -68,13 +107,14 @@ pub trait ViewportMutable: RawElWrapper {
         let limit_to_body = limit_to_body.into();
         self.update_raw_el(move |raw_el| {
             raw_el
-                .insert(MutableViewportSettings { limit_to_body })
+                .insert(MutableViewport::new(limit_to_body))
+                // .observe(observer)
                 .observe(
                     move |mutation: Trigger<ViewportMutation>,
                           mut styles: Query<&mut Style>,
                           parents: Query<&Parent>,
                           nodes: Query<&Node>,
-                          settings: Query<&MutableViewportSettings>| {
+                          settings: Query<&MutableViewport>| {
                         let entity = mutation.entity();
                         if let Some((((node, parent), settings), mut style)) = nodes
                             .get(entity)
@@ -109,13 +149,13 @@ pub trait ViewportMutable: RawElWrapper {
                     },
                 )
                 .defer_update(DeferredUpdaterAppendDirection::Front, move |raw_el| {
-                    // this wrapper element is the [`Viewport`]
                     RawHaalkaEl::from(NodeBundle::default())
+                        .insert(ViewportMarker)
                         .with_component::<Style>(move |mut style| {
                             style.display = Display::Flex;
                             style.overflow = overflow;
                         })
-                        .child(raw_el) // the `raw_el` here is the [`Scene`]
+                        .child(raw_el) // this is the [`Scene`]
                         .on_spawn_with_system(
                             |In(entity), children: Query<&Children>, mut styles: Query<&mut Style>| {
                                 // match the flex direction of `raw_el` above
@@ -137,8 +177,29 @@ pub trait ViewportMutable: RawElWrapper {
         })
     }
 
-    // TODO
-    // fn on_viewport_location_change(self, mut handler: impl FnMut(Scene, Viewport) + 'static) -> Self
+    fn on_viewport_location_change_with_system<Marker>(
+        self,
+        handler: impl IntoSystem<(Entity, (Scene, Viewport)), (), Marker> + Send + 'static,
+    ) -> Self {
+        self.update_raw_el(|raw_el| {
+            let system_holder = Mutable::new(None);
+            raw_el
+            .insert(OnViewportLocationChange)
+            .on_spawn(clone!((system_holder) move |world, entity| {
+                let system = register_system(world, handler);
+                system_holder.set(Some(system));
+                observe(world, entity, move |viewport_location_change: Trigger<ViewportLocationChange>, mut commands: Commands| {
+                    let &ViewportLocationChange { scene, viewport } = viewport_location_change.event();
+                    commands.run_system_with_input(system, (entity, (scene, viewport)));
+                });
+            }))
+            .apply(remove_system_holder_on_remove(system_holder))
+        })
+    }
+
+    fn on_viewport_location_change(self, mut handler: impl FnMut(Scene, Viewport) + Send + Sync + 'static) -> Self {
+        self.on_viewport_location_change_with_system(move |In((_, (scene, viewport)))| handler(scene, viewport))
+    }
 
     /// Reactively set the horizontal position of the viewport.
     fn viewport_x_signal<S: Signal<Item = f32> + Send + 'static>(
@@ -175,4 +236,59 @@ pub trait ViewportMutable: RawElWrapper {
         }
         self
     }
+}
+
+#[derive(Event)]
+struct ViewportLocationChange {
+    scene: Scene,
+    viewport: Viewport,
+}
+
+fn scene_change_dispatcher(
+    mut data: Query<(Entity, &Node, &Style, &mut MutableViewport), Or<(Changed<Node>, Changed<Transform>)>>,
+    mut commands: Commands,
+) {
+    for (entity, node, style, mut mutable_viewport) in data.iter_mut() {
+        let Vec2 { x, y } = node.size();
+        mutable_viewport.scene.width = x;
+        mutable_viewport.scene.height = y;
+        if let Val::Px(x) = style.left {
+            mutable_viewport.viewport.x = -x;
+        }
+        if let Val::Px(y) = style.top {
+            mutable_viewport.viewport.y = -y;
+        }
+        let MutableViewport { scene, viewport, .. } = *mutable_viewport;
+        commands.trigger_targets(ViewportLocationChange { scene, viewport }, entity);
+    }
+}
+
+fn viewport_change_dispatcher(
+    data: Query<(Entity, &Node), (With<ViewportMarker>, Changed<Node>)>,
+    children: Query<&Children>,
+    mut mutable_viewports: Query<&mut MutableViewport>,
+    mut commands: Commands,
+) {
+    for (entity, node) in data.iter() {
+        let Vec2 { x, y } = node.size();
+        if let Ok(children) = children.get(entity) {
+            // [`Scene`] is the [`Viewport`]'s only child
+            if let Some(&child) = children.first() {
+                if let Ok(mut mutable_viewport) = mutable_viewports.get_mut(child) {
+                    mutable_viewport.viewport.width = x;
+                    mutable_viewport.viewport.height = y;
+                    let MutableViewport { scene, viewport, .. } = *mutable_viewport;
+                    commands.trigger_targets(ViewportLocationChange { scene, viewport }, child);
+                }
+            }
+        }
+    }
+}
+
+pub(super) fn plugin(app: &mut App) {
+    app.add_systems(
+        Update,
+        (scene_change_dispatcher, viewport_change_dispatcher)
+            .run_if(any_with_component::<MutableViewport>.and_then(any_with_component::<OnViewportLocationChange>)),
+    );
 }